feat: organização do objeto de tema
