+++ client/src/styles/themes/default.ts
@@ -1,0 +1,24 @@
+export const defaultTheme = {
+  colors: {
+    white: '#FFFFFF',
+
+    soft_gray: '#F5F5F5',
+    gray: '#B9B6BE',
+    dark_gray: '#AAAAAA',
+    black: '#000000',
+
+    soft_puple: '#BA98E6',
+    puple: '#8C53D6',
+    dark_puple: '#574969',
+
+    cyane: '#B6E3F4',
+  },
+  typography: {
+    fontSizes: {
+      xs: '8px',
+      sm: '12px',
+      md: '18px',
+      lg: '24px',
+    },
+  },
+}
