--- client/src/themes/default.ts
@@ -1,13 +1,0 @@
-export const defaultTheme = {
-  white: '#FFFFFF',
-  'soft-gray': '#F5F5F5',
-  gray: '#B9B6BE',
-  'dark-gray': '#AAAAAA',
-  black: '#000000',
-
-  'soft-puple': '#BA98E6',
-  puple: '#8C53D6',
-  'dark-puple': '#574969',
-
-  cyane: '#B6E3F4',
-}
